requirements.txt
streamlit
python-dotenv

openai
requests

langchain
langchain-openai
langchain-community
langchain-text-splitters
langchain-tavily
langgraph

faiss-cpu
pypdf
